+++ src/testing/mod.rs
@@ -1,0 +1,5 @@
+//! Test utilities.
+
+mod fake_server;
+
+pub use self::fake_server::FakeServer;